changes messages going to hardware controller

+++ train_model/to_TC.py
@@ -21,8 +21,8 @@ class to_TC(msg):
 		self.signalPickup_failure = None
 		self.__vars__.append(self.signalPickup_failure)
 		self.__data_types__.append("float")
-		self.__data_types__.append("int")
-		self.__data_types__.append("int")
+		self.__data_types__.append("float")
+		self.__data_types__.append("float")
 		self.__data_types__.append("float")
 		self.__data_types__.append("string")
 		self.__data_types__.append("bool")
